Propagate masksed bases
In some cases, marginal matches to repetitive elements were getting
masked in some species but not others. So the new plan is to do
alignments with all bases included, then mask sequences, then propagate
the masked base to the alignement.

If a base gets masked, then all bases that it aligns to also get masked.